+++ packages/cli/src/utils/get-command-name-from-args.ts
